Dependencias modificadas. Mapeo correcto de clase Cliente

+++ src/main/java/com/fra/clientes/dao/ClienteDao.java
@@ -8,7 +8,7 @@ public interface ClienteDao {
 
 	public void addCliente(Cliente c);
 	public void updateCliente(Cliente c);
-	public void removeCliente(int id);
-	public Cliente getClienteById(int id);
+	public void removeCliente(long id);
+	public Cliente getClienteById(long id);
 	public List<Cliente> getClientes();
 }
